Move gql models to proper package
