+++ src/ai/backend/manager/services/domain/types.py
@@ -8,7 +8,7 @@ from sqlalchemy.engine.result import Row
 from ai.backend.common.types import ResourceSlot, VFolderHostPermissionMap
 from ai.backend.manager.models.domain import DomainRow
 from ai.backend.manager.models.user import UserRole
-from ai.backend.manager.types import Creator, OptionalState, PartialModifier, State, TriState
+from ai.backend.manager.types import Creator, OptionalState, PartialModifier, TriState, TriStateEnum
 
 
 @dataclass
@@ -107,19 +107,19 @@ class DomainModifier(PartialModifier):
     @override
     def get_modified_fields(self) -> dict[str, Any]:
         modified: dict[str, Any] = {}
-        if self.name.state() != State.NOP:
+        if self.name.state() != TriStateEnum.NOP:
             modified["name"] = self.name.value()
-        if self.description.state() != State.NOP:
+        if self.description.state() != TriStateEnum.NOP:
             modified["description"] = self.description.value()
-        if self.is_active.state() != State.NOP:
+        if self.is_active.state() != TriStateEnum.NOP:
             modified["is_active"] = self.is_active.value()
-        if self.total_resource_slots.state() != State.NOP:
+        if self.total_resource_slots.state() != TriStateEnum.NOP:
             modified["total_resource_slots"] = self.total_resource_slots.value()
-        if self.allowed_vfolder_hosts.state() != State.NOP:
+        if self.allowed_vfolder_hosts.state() != TriStateEnum.NOP:
             modified["allowed_vfolder_hosts"] = self.allowed_vfolder_hosts.value()
-        if self.allowed_docker_registries.state() != State.NOP:
+        if self.allowed_docker_registries.state() != TriStateEnum.NOP:
             modified["allowed_docker_registries"] = self.allowed_docker_registries.value()
-        if self.integration_id.state() != State.NOP:
+        if self.integration_id.state() != TriStateEnum.NOP:
             modified["integration_id"] = self.integration_id.value()
         return modified
 
@@ -147,18 +147,18 @@ class DomainNodeModifier(PartialModifier):
     @override
     def get_modified_fields(self) -> dict[str, Any]:
         modified: dict[str, Any] = {}
-        if self.description.state() != State.NOP:
+        if self.description.state() != TriStateEnum.NOP:
             modified["description"] = self.description.value()
-        if self.is_active.state() != State.NOP:
+        if self.is_active.state() != TriStateEnum.NOP:
             modified["is_active"] = self.is_active.value()
-        if self.total_resource_slots.state() != State.NOP:
+        if self.total_resource_slots.state() != TriStateEnum.NOP:
             modified["total_resource_slots"] = self.total_resource_slots.value()
-        if self.allowed_vfolder_hosts.state() != State.NOP:
+        if self.allowed_vfolder_hosts.state() != TriStateEnum.NOP:
             modified["allowed_vfolder_hosts"] = self.allowed_vfolder_hosts.value()
-        if self.allowed_docker_registries.state() != State.NOP:
+        if self.allowed_docker_registries.state() != TriStateEnum.NOP:
             modified["allowed_docker_registries"] = self.allowed_docker_registries.value()
-        if self.integration_id.state() != State.NOP:
+        if self.integration_id.state() != TriStateEnum.NOP:
             modified["integration_id"] = self.integration_id.value()
-        if self.dotfiles.state() != State.NOP:
+        if self.dotfiles.state() != TriStateEnum.NOP:
             modified["dotfiles"] = self.dotfiles.value()
         return modified
